added read functions moved classes to their own files.
