+++ FF8/FF8.csproj
@@ -153,20 +153,47 @@
     <Compile Include="Menu\Kernel_bin.Attack_Flags.cs" />
     <Compile Include="Menu\Kernel_bin.Battle_Commands.cs" />
     <Compile Include="Menu\Kernel_bin.Battle_Items.cs" />
+    <Compile Include="Menu\Kernel_bin.Blue_magic_Quistis_limit_break.cs" />
+    <Compile Include="Menu\Kernel_bin.Character_abilities.cs" />
     <Compile Include="Menu\Kernel_bin.Character_Stats.cs" />
+    <Compile Include="Menu\Kernel_bin.Command_abilities.cs" />
+    <Compile Include="Menu\Kernel_bin.Command_ability.cs" />
+    <Compile Include="Menu\Kernel_bin.Command_ability_data.cs" />
     <Compile Include="Menu\Kernel_bin.cs" />
+    <Compile Include="Menu\Kernel_bin.Devour.cs" />
+    <Compile Include="Menu\Kernel_bin.Duel_Zell_limit_break.cs" />
     <Compile Include="Menu\Kernel_bin.Element.cs" />
     <Compile Include="Menu\Kernel_bin.Enemy_Attacks_Data.cs" />
     <Compile Include="Menu\Kernel_bin.Gender.cs" />
+    <Compile Include="Menu\Kernel_bin.GF_abilities.cs" />
     <Compile Include="Menu\Kernel_bin.Junctionable_GFs_Data.cs" />
+    <Compile Include="Menu\Kernel_bin.Junction_abilities.cs" />
     <Compile Include="Menu\Kernel_bin.Magic_Data.cs" />
     <Compile Include="Menu\Kernel_bin.Magic_ID.cs" />
+    <Compile Include="Menu\Kernel_bin.Menu_abilities.cs" />
+    <Compile Include="Menu\Kernel_bin.Misc_section.cs" />
+    <Compile Include="Menu\Kernel_bin.Misc_text_pointers.cs" />
+    <Compile Include="Menu\Kernel_bin.Non_battle_Items_Data.cs" />
     <Compile Include="Menu\Kernel_bin.Non_Junctionable_GFs_Attacks_Data.cs" />
+    <Compile Include="Menu\Kernel_bin.Party_abilities.cs" />
+    <Compile Include="Menu\Kernel_bin.Quanity.cs" />
+    <Compile Include="Menu\Kernel_bin.Quistis_limit_break_parameters.cs" />
     <Compile Include="Menu\Kernel_bin.Renzokuken_Finishers_Data.cs" />
+    <Compile Include="Menu\Kernel_bin.Rinoa_limit_breaks_part_1.cs" />
+    <Compile Include="Menu\Kernel_bin.Rinoa_limit_breaks_part_2.cs" />
+    <Compile Include="Menu\Kernel_bin.Selphie_limit_break_sets.cs" />
+    <Compile Include="Menu\Kernel_bin.Shot_Irvine_limit_break.cs" />
+    <Compile Include="Menu\Kernel_bin.Slot.cs" />
+    <Compile Include="Menu\Kernel_bin.Slot_array.cs" />
+    <Compile Include="Menu\Kernel_bin.Stat.cs" />
+    <Compile Include="Menu\Kernel_bin.StatFlags.cs" />
     <Compile Include="Menu\Kernel_bin.Statuses0.cs" />
     <Compile Include="Menu\Kernel_bin.Statuses1.cs" />
+    <Compile Include="Menu\Kernel_bin.Stat_percent_abilities.cs" />
     <Compile Include="Menu\Kernel_bin.Target.cs" />
+    <Compile Include="Menu\Kernel_bin.Temporary_character_limit_breaks.cs" />
     <Compile Include="Menu\Kernel_bin.Weapons_Data.cs" />
+    <Compile Include="Menu\Kernel_bin.Zell_limit_break_parameters.cs" />
     <Compile Include="Menu\Magazine.cs" />
     <Compile Include="Menu\Magazine.InsertCustomEntries.cs" />
     <Compile Include="Menu\Magazine.ID.cs" />
