+++ FF8/Menu/Kernel_bin.Command_ability.cs
@@ -1,0 +1,25 @@
+﻿namespace FF8
+{
+    internal partial class Kernel_bin
+    {
+        /// <summary>
+        /// Command Ability
+        /// </summary>
+        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Command-ability-data"/>
+        internal enum Command_ability
+        {
+            Recover,
+            Revive,
+            Treatment,
+            Mad_Rush,
+            Doom,
+            Absorb,
+            LV_Down,
+            LV_Up,
+            Kamikaze,
+            Devour,
+            Card,
+            Defend,
+        }
+    }
+}
