+++ FF8/Menu/Kernel_bin.Rinoa_limit_breaks_part_1.cs
@@ -1,0 +1,57 @@
+﻿using System.Collections;
+using System.IO;
+
+namespace FF8
+{
+    internal partial class Kernel_bin
+    {
+        /// <summary>
+        /// Rinoa limit breaks (part 1)
+        /// </summary>
+        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Rinoa-limit-breaks-%28part-1%29"/>
+        internal class Rinoa_limit_breaks_part_1
+        {
+            internal const int count = 2;
+            internal const int id = 24;
+            internal const int size = 8;
+
+            public override string ToString() => Name;
+
+            public FF8String Name { get; private set; }
+            public FF8String Description { get; private set; }
+            public BitArray Unknown0 { get; private set; }
+            public Target Target { get; private set; }
+            public byte AbilityID { get; private set; }
+            public byte Unknown1 { get; private set; }
+
+            internal void Read(BinaryReader br, int i)
+            {
+                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
+                //0x0000	2 bytes Offset to name
+                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
+                //0x0002	2 bytes Offset to description
+                br.BaseStream.Seek(4, SeekOrigin.Current);
+                Unknown0 = new BitArray(br.ReadBytes(1));
+                //0x0004  1 byte Unknown Flags
+                Target = (Target)br.ReadByte();
+                //0x0005  1 byte Target
+                AbilityID = br.ReadByte();
+                //0x0006  1 byte Ability data ID
+                Unknown1 = br.ReadByte();
+                //0x0007  1 byte Unknown / Unused
+            }
+            internal static Rinoa_limit_breaks_part_1[] Read(BinaryReader br)
+            {
+                var ret = new Rinoa_limit_breaks_part_1[count];
+
+                for (int i = 0; i < count; i++)
+                {
+                    var tmp = new Rinoa_limit_breaks_part_1();
+                    tmp.Read(br, i);
+                    ret[i] = tmp;
+                }
+                return ret;
+            }
+        }
+    }
+}
