+++ FF8/Menu/Kernel_bin.Stat.cs
@@ -1,0 +1,18 @@
+﻿namespace FF8
+{
+    internal partial class Kernel_bin
+    {
+        internal enum Stat
+        {
+            HP,
+            STR,
+            VIT,
+            MAG,
+            SPR,
+            SPD,
+            EVA,
+            HIT,
+            LUCK
+        }
+    }
+}
