+++ FF8/Menu/Kernel_bin.Zell_limit_break_parameters.cs
@@ -1,0 +1,35 @@
+﻿using System.IO;
+
+namespace FF8
+{
+    internal partial class Kernel_bin
+    {
+        /// <summary>
+        /// Zell Limit Break Parameters
+        /// </summary>
+        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Zell-limit-break-parameters"/>
+        internal class Zell_limit_break_parameters
+        {
+            internal const int count = 24;
+            internal const int id = 23;
+            internal const int size = 4;
+
+            public byte[] Moves { get; private set; }
+
+            internal void Read(BinaryReader br, int i) => Moves = br.ReadBytes(4);//0x0000  1 byte[[StartMove 0//0x0001  1 byte  Next Sequence 0_1//0x0002  1 byte  Next Sequence 0_2//0x0003  1 byte  Next Sequence 0_3
+
+            internal static Zell_limit_break_parameters[] Read(BinaryReader br)
+            {
+                var ret = new Zell_limit_break_parameters[count];
+
+                for (int i = 0; i < count; i++)
+                {
+                    var tmp = new Zell_limit_break_parameters();
+                    tmp.Read(br, i);
+                    ret[i] = tmp;
+                }
+                return ret;
+            }
+        }
+    }
+}
