+++ FF8/Menu/Kernel_bin.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
 
 namespace FF8
@@ -138,943 +136,50 @@ namespace FF8
                 {
                     NonJunctionableGFsAttacksData[i].Read(br, i);
                 }
+
+                ms.Seek(subPositions[Command_ability_data.id], SeekOrigin.Begin);
+                Commandabilitydata = Command_ability_data.Read(br);
+                ms.Seek(subPositions[Junction_abilities.id], SeekOrigin.Begin);
+                Junctionabilities = Junction_abilities.Read(br);
+                ms.Seek(subPositions[Command_abilities.id], SeekOrigin.Begin);
+                Commandabilities = Command_abilities.Read(br);
+                ms.Seek(subPositions[Stat_percent_abilities.id], SeekOrigin.Begin);
+                Statpercentabilities = Stat_percent_abilities.Read(br);
+                ms.Seek(subPositions[Character_abilities.id], SeekOrigin.Begin);
+                Characterabilities = Character_abilities.Read(br);
+                ms.Seek(subPositions[Party_abilities.id], SeekOrigin.Begin);
+                Partyabilities = Party_abilities.Read(br);
+                ms.Seek(subPositions[GF_abilities.id], SeekOrigin.Begin);
+                GFabilities = GF_abilities.Read(br);
+                ms.Seek(subPositions[Menu_abilities.id], SeekOrigin.Begin);
+                Menuabilities = Menu_abilities.Read(br);
+                ms.Seek(subPositions[Temporary_character_limit_breaks.id], SeekOrigin.Begin);
+                Temporarycharacterlimitbreaks = Temporary_character_limit_breaks.Read(br);
+                ms.Seek(subPositions[Blue_magic_Quistis_limit_break.id], SeekOrigin.Begin);
+                BluemagicQuistislimitbreak = Blue_magic_Quistis_limit_break.Read(br);
+                ms.Seek(subPositions[Quistis_limit_break_parameters.id], SeekOrigin.Begin);
+                Quistislimitbreakparameters = Quistis_limit_break_parameters.Read(br);
+                ms.Seek(subPositions[Shot_Irvine_limit_break.id], SeekOrigin.Begin);
+                ShotIrvinelimitbreak = Shot_Irvine_limit_break.Read(br);
+                ms.Seek(subPositions[Duel_Zell_limit_break.id], SeekOrigin.Begin);
+                DuelZelllimitbreak = Duel_Zell_limit_break.Read(br);
+                ms.Seek(subPositions[Zell_limit_break_parameters.id], SeekOrigin.Begin);
+                Zelllimitbreakparameters = Zell_limit_break_parameters.Read(br);
+                ms.Seek(subPositions[Rinoa_limit_breaks_part_1.id], SeekOrigin.Begin);
+                Rinoalimitbreakspart1 = Rinoa_limit_breaks_part_1.Read(br);
+                ms.Seek(subPositions[Rinoa_limit_breaks_part_2.id], SeekOrigin.Begin);
+                Rinoalimitbreakspart2 = Rinoa_limit_breaks_part_2.Read(br);
+                ms.Seek(subPositions[Slot_array.id], SeekOrigin.Begin);
+                Slotarray = Slot_array.Read(br);
+                ms.Seek(subPositions[Selphie_limit_break_sets.id], SeekOrigin.Begin);
+                Selphielimitbreaksets = Selphie_limit_break_sets.Read(br);
+                ms.Seek(subPositions[Devour.id], SeekOrigin.Begin);
+                Devour_ = Devour.Read(br);
+                ms.Seek(subPositions[Misc_section.id], SeekOrigin.Begin);
+                Miscsection = Misc_section.Read(br);
+                ms.Seek(subPositions[Misc_text_pointers.id], SeekOrigin.Begin);
+                Misctextpointers = Misc_text_pointers.Read();
             }
-        }
-
-        /// <summary>
-        /// Non battle Items Mame and Description Offsets Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Non-battle-item-name-and-description-offsets"/>
-        internal struct Non_battle_Items_Data
-        {
-            internal static readonly int count = 166;
-            internal static readonly int id = 8;
-
-            public override string ToString() => Name;
-
-            internal FF8String Name { get; private set; }
-
-            //0x0000	2 bytes Offset to item name
-            internal FF8String Description { get; private set; }
-
-            //0x0002	2 bytes Offset to item description
-
-            internal void Read(int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to item name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to item description
-                //br.BaseStream.Seek(4,SeekOrigin.Current);
-            }
-        }
-
-        /// <summary>
-        /// Command Ability
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Command-ability-data"/>
-        internal enum Command_ability
-        {
-            Recover,
-            Revive,
-            Treatment,
-            Mad_Rush,
-            Doom,
-            Absorb,
-            LV_Down,
-            LV_Up,
-            Kamikaze,
-            Devour,
-            Card,
-            Defend,
-        }
-
-        /// <summary>
-        /// Command Abilities Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Command-ability-data"/>
-        internal class Command_ability_data
-        {
-            internal const int count = 12;
-            internal const int id = 10;
-
-            public Magic_ID MagicID { get; private set; }
-            public byte[] Unknown0 { get; private set; }
-            public Attack_Type Attack_Type { get; private set; }
-            public byte Attack_Power { get; private set; }
-            public Attack_Flags Attack_Flags { get; private set; }
-            public byte Hit_Count { get; private set; }
-            public Element Element { get; private set; }
-            public byte Status_Attack { get; private set; }
-            public Statuses0 Statuses0 { get; private set; }
-            public Statuses1 Statuses1 { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                MagicID = (Magic_ID)br.ReadUInt16();
-                //0x0000  2 bytes Magic ID
-                Unknown0 = br.ReadBytes(2);
-                //0x0002  2 bytes Unknown
-                Attack_Type = (Attack_Type)br.ReadByte();
-                //0x0004  1 byte Attack type
-                Attack_Power = br.ReadByte();
-                //0x0005  1 byte Attack power
-                Attack_Flags = (Attack_Flags)br.ReadByte();
-                //0x0006  1 byte Attack flags
-                Hit_Count = br.ReadByte();
-                //0x0007  1 byte Hit Count
-                Element = (Element)br.ReadByte();
-                //0x0008  1 byte Element
-                Status_Attack = br.ReadByte();
-                //0x0009  1 byte Status attack enabler
-                Statuses0 = (Statuses0)br.ReadUInt16();
-                Statuses1 = (Statuses1)br.ReadUInt32();
-                //0x000A  6 bytes Statuses
-            }
-        }
-
-        /// <summary>
-        /// Junction Abilities Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Junction-abilities"/>
-        internal class Junction_abilities
-        {
-            internal const int count = 20;
-            internal const int id = 11;
-
-            public override string ToString() => Name;
-
-            internal FF8String Name { get; private set; }
-            internal FF8String Description { get; private set; }
-            public byte AP { get; private set; }
-            public BitArray J_Flags { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-                br.BaseStream.Seek(4, SeekOrigin.Current);
-                AP = br.ReadByte();
-                //0x0004  1 byte AP Required to learn ability
-                J_Flags = new BitArray(br.ReadBytes(3));
-                //0x0005  3 byte J-Flag
-            }
-        }
-
-        /// <summary>
-        /// Command Abilities
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Command-abilities"/>
-        internal class Command_abilities
-        {
-            internal const int count = 19;
-            internal const int id = 12;
-
-            public override string ToString() => Name;
-
-            internal FF8String Name { get; private set; }
-            internal FF8String Description { get; private set; }
-            public byte AP { get; private set; }
-            public byte Index { get; private set; }
-            public byte[] Unknown0 { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-
-                AP = br.ReadByte();
-                //0x0004  1 byte AP Required to learn ability
-                Index = br.ReadByte();
-                //0x0005  1 byte Index to Battle commands
-                Unknown0 = br.ReadBytes(2);
-                //0x0006  2 bytes Unknown/ Unused
-            }
-        }
-
-        /// <summary>
-        /// Stat Percentage Increasing Abilities Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Stat-percentage-increasing-abilities"/>
-        internal class Stat_percent_abilities
-        {
-            internal const int count = 19;
-            internal const int id = 13;
-
-            public override string ToString() => Name;
-            public FF8String Name { get; private set; }
-            public FF8String Description { get; private set; }
-            public byte AP { get; private set; }
-            public Stat Stat { get; private set; }
-            public byte Value { get; private set; }
-            public byte Unknown0 { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-                AP = br.ReadByte();
-                //0x0004  1 byte AP needed to learn the ability
-                Stat = (Stat)br.ReadByte();
-                //0x0005  1 byte Stat to increase
-                Value = br.ReadByte();
-                //0x0006  1 byte Increase value
-                Unknown0 = br.ReadByte();
-                //0x0007  1 byte Unknown/ Unused
-            }
-        }
-
-        /// <summary>
-        /// Characters Abilities Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Character-abilities"/>
-        internal class Character_abilities
-        {
-            internal const int count = 20;
-            internal const int id = 14;
-
-            public override string ToString() => Name;
-            public FF8String Name { get; private set; }
-            public FF8String Description { get; private set; }
-            public byte AP { get; private set; }
-            public BitArray Flags { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-                br.BaseStream.Seek(4, SeekOrigin.Current);
-                AP = br.ReadByte();
-                //0x0004  1 byte AP Required to learn ability
-                Flags = new BitArray(br.ReadBytes(3));
-                //0x0005  3 byte Flags
-            }
-        }
-
-        /// <summary>
-        /// Party Abilities Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Party-abilities"/>
-        internal class Party_abilities
-        {
-            internal const int count = 5;
-            internal const int id = 15;
-
-            public override string ToString() => Name;
-            public FF8String Name { get; private set; }
-            public FF8String Description { get; private set; }
-            public byte AP { get; private set; }
-            public BitArray Flags { get; private set; }
-            public byte[] Unknown0 { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-                br.BaseStream.Seek(4, SeekOrigin.Current);
-                AP = br.ReadByte();
-                //0x0004  1 byte AP Required to learn ability
-                Flags = new BitArray(br.ReadBytes(1));
-                //0x0005  1 byte Flags
-                Unknown0 = br.ReadBytes(2);
-                //0x0006  2 byte Unknown/ Unused
-            }
-        }
-
-        internal enum Stat
-        {
-            HP,
-            STR,
-            VIT,
-            MAG,
-            SPR,
-            SPD,
-            EVA,
-            HIT,
-            LUCK
-        }
-
-        /// <summary>
-        /// GF Abilities Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/GF-abilities"/>
-        internal class GF_abilities
-        {
-            internal const int count = 9;
-            internal const int id = 16;
-
-            public override string ToString() => Name;
-            public FF8String Name { get; private set; }
-            public FF8String Description { get; private set; }
-            public byte AP { get; private set; }
-            public byte Boost { get; private set; }
-            public Stat Stat { get; private set; }
-            public byte Value { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-                br.BaseStream.Seek(4, SeekOrigin.Current);
-                AP = br.ReadByte();
-                //0x0004 1 byte AP needed to learn the ability
-                Boost = br.ReadByte();
-                //0x0005 Enable Boost
-                Stat = (Stat)br.ReadByte();
-                //0x0006  1 byte Stat to increase
-                Value = br.ReadByte();
-                //0x0007  1 byte Increase value
-            }
-        }
-
-        /// <summary>
-        /// Menu Abilities Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Menu-abilities"/>
-        internal class Menu_abilities
-        {
-            internal const int count = 24;
-            internal const int id = 17;
-
-            public override string ToString() => Name;
-            public FF8String Name { get; private set; }
-            public FF8String Description { get; private set; }
-            public byte AP { get; private set; }
-            public byte Index { get; private set; }
-            public byte Start { get; private set; }
-            public byte End { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-                br.BaseStream.Seek(4, SeekOrigin.Current);
-                AP = br.ReadByte();
-                //0x0004  1 byte AP Required to learn ability
-                Index = br.ReadByte();
-                //0x0005  1 byte Index to m00X files in menu.fs
-                //(first 3 sections are treated as special cases)
-                Start = br.ReadByte();
-                //0x0006  1 byte Start offset
-                End = br.ReadByte();
-                //0x0007  1 byte End offset
-            }
-        }
-
-        /// <summary>
-        /// Temporary Characters Limit Breaks
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Temporary-character-limit-breaks"/>
-        internal class Temporary_character_limit_breaks
-        {
-            internal const int count = 5;
-            internal const int id = 18;
-
-            public override string ToString() => Name;
-            public FF8String Name { get; private set; }
-            public FF8String Description { get; private set; }
-            public Magic_ID MagicID { get; private set; }
-            public Attack_Type Attack_Type { get; private set; }
-            public byte Attack_Power { get; private set; }
-            public byte[] Unknown0 { get; private set; }
-            public Target Target { get; private set; }
-            public Attack_Flags Attack_Flags { get; private set; }
-            public byte Hit_Count { get; private set; }
-            public Element Element { get; private set; }
-            public byte Element_Percent { get; private set; }
-            public byte Status_Attack { get; private set; }
-            public Statuses0 Statuses0 { get; private set; }
-            public byte[] Unknown1 { get; private set; }
-            public Statuses1 Statuses1 { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-                br.BaseStream.Seek(4, SeekOrigin.Current);
-                MagicID = (Magic_ID)br.ReadUInt16();
-                //0x0004  2 bytes Magic ID
-                Attack_Type = (Attack_Type)br.ReadByte();
-                //0x0006  1 byte Attack Type
-                Attack_Power = br.ReadByte();
-                //0x0007  1 byte Attack Power
-                Unknown0 = br.ReadBytes(2);
-                //0x0008  2 bytes Unknown
-                Target = (Target)br.ReadByte();
-                //0x000A  1 byte Target Info
-                Attack_Flags = (Attack_Flags)br.ReadByte();
-                //0x000B  1 byte Attack Flags
-                Hit_Count = br.ReadByte();
-                //0x000C  1 byte Hit Count
-                Element = (Element)br.ReadByte();
-                //0x000D  1 byte Element Attack
-                Element_Percent = br.ReadByte();
-                //0x000E  1 byte Element Attack %
-                Status_Attack = br.ReadByte();
-                //0x000F  1 byte Status Attack Enabler
-                Statuses0 = (Statuses0)br.ReadUInt16();
-                //0x0010  2 bytes status_0; //statuses 0-7
-                Unknown1 = br.ReadBytes(2);
-                //0x0012  2 bytes Unknown
-                Statuses1 = (Statuses1)br.ReadUInt32();
-                //0x0014  4 bytes status_1; //statuses 8-39
-            }
-        }
-
-        /// <summary>
-        /// Blue magic (Quistis limit break)
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Blue-magic-%28Quistis-limit-break%29"/>
-        internal class Blue_magic_Quistis_limit_break
-        {
-            internal const int count = 16;
-            internal const int id = 19;
-
-            public override string ToString() => Name;
-            public FF8String Name { get; private set; }
-            public FF8String Description { get; private set; }
-            public Magic_ID MagicID { get; private set; }
-            public byte Unknown0 { get; private set; }
-            public Attack_Type Attack_Type { get; private set; }
-            public byte[] Unknown1 { get; private set; }
-            public Attack_Flags Attack_Flags { get; private set; }
-            public byte Unknown2 { get; private set; }
-            public Element Element { get; private set; }
-            public byte Status_Attack { get; private set; }
-            public byte Crit { get; private set; }
-            public byte Unknown3 { get; private set; }
-            public Quistis_limit_break_parameters[] Crisis_Levels { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-                br.BaseStream.Seek(4, SeekOrigin.Current);
-                MagicID = (Magic_ID)br.ReadUInt16();
-                //0x0004  2 bytes Magic ID
-                Unknown0 = br.ReadByte();
-                //0x0006  1 byte Unknown
-                Attack_Type = (Attack_Type)br.ReadByte();
-                //0x0007  1 byte Attack Type
-                Unknown1 = br.ReadBytes(2);
-                //0x0008  2 bytes Unknown
-                Attack_Flags = (Attack_Flags)br.ReadByte();
-                //0x000A  1 byte Attack Flags
-                Unknown2 = br.ReadByte();
-                //0x000B  1 byte Unknown
-                Element = (Element)br.ReadByte();
-                //0x000C  1 byte Element
-                Status_Attack = br.ReadByte();
-                //0x000D  1 byte Status Attack
-                Crit = br.ReadByte();
-                //0x000E  1 byte Crit Bonus
-                Unknown3 = br.ReadByte();
-                //0x000F  1 byte Unknown
-                long current = br.BaseStream.Position;
-
-                br.BaseStream.Seek(Memory.Strings.Files[Strings.FileID.KERNEL].subPositions[Quistis_limit_break_parameters.id + Quistis_limit_break_parameters.size * i], SeekOrigin.Begin);
-                Crisis_Levels = new Quistis_limit_break_parameters[Quistis_limit_break_parameters.size];
-                for (i = 0; i < Quistis_limit_break_parameters.count; i++)
-                    Crisis_Levels[i].Read(br, i);
-                br.BaseStream.Seek(current, SeekOrigin.Begin);
-            }
-        }
-
-        /// <summary>
-        /// Blue Magic Parameters - 4 for each spell for crisis level.
-        /// </summary>
-        /// <see cref="https://finalfantasy.fandom.com/wiki/Blue_Magic_(Final_Fantasy_VIII)"/>
-        /// <seealso cref="https://github.com/alexfilth/doomtrain/wiki/Quistis-limit-break-parameters"/>
-        internal struct Quistis_limit_break_parameters
-        {
-            internal const int count = 4;//64 total but I want to add these to the Blue_magic_Quistis_limit_break in an array
-            internal const int id = 20;
-            internal const int size = 8;
-
-            public Statuses1 Statuses1 { get; private set; }
-            public Statuses0 Statuses0 { get; private set; }
-            public byte Attack_Power { get; private set; }
-            public byte Attack_Param { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Statuses1 = (Statuses1)br.ReadUInt32();
-                //0x0000  4 bytes Status 1
-                Statuses0 = (Statuses0)br.ReadUInt16();
-                //0x0004  2 bytes Status 0
-                Attack_Power = br.ReadByte();
-                //0x0006  1 bytes Attack Power
-                Attack_Param = br.ReadByte();
-                //0x0007  1 byte Attack Param
-            }
-        }
-
-        /// <summary>
-        /// Shot (Irvine limit break)
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Shot-%28Irvine-limit-break%29"/>
-        internal class Shot_Irvine_limit_break
-        {
-            internal const int count = 8;
-            internal const int id = 21;
-            internal const int size = 24;
-
-            public override string ToString() => Name;
-            public FF8String Name { get; private set; }
-            public FF8String Description { get; private set; }
-            public Magic_ID MagicID { get; private set; }
-            public Attack_Type Attack_Type { get; private set; }
-            public byte Attack_Power { get; private set; }
-            public byte[] Unknown0 { get; private set; }
-            public Target Target { get; private set; }
-            public Attack_Flags Attack_Flags { get; private set; }
-            public byte Hit_Count { get; private set; }
-            public Element Element { get; private set; }
-            public byte Element_Percent { get; private set; }
-            public byte Status_Attack { get; private set; }
-            public Statuses0 Statuses0 { get; private set; }
-            public byte Item_Index { get; private set; }
-            public byte Crit { get; private set; }
-            public Statuses1 Statuses1 { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-                br.BaseStream.Seek(4, SeekOrigin.Current);
-                MagicID = (Magic_ID)br.ReadUInt16();
-                //0x0004  2 bytes Magic ID
-                Attack_Type = (Attack_Type)br.ReadByte();
-                //0x0006  1 byte Attack Type
-                Attack_Power = br.ReadByte();
-                //0x0007  1 byte Attack Power
-                Unknown0 = br.ReadBytes(2);
-                //0x0008  2 bytes Unknown
-                Target = (Target)br.ReadByte();
-                //0x000A  1 byte Target Info
-                Attack_Flags = (Attack_Flags)br.ReadByte();
-                //0x000B  1 byte Attack Flags
-                Hit_Count = br.ReadByte();
-                //0x000C  1 byte Hit Count
-                Element = (Element)br.ReadByte();
-                //0x000D  1 byte Element Attack
-                Element_Percent = br.ReadByte();
-                //0x000E  1 byte Element Attack %
-                Status_Attack = br.ReadByte();
-                //0x000F  1 byte Status Attack Enabler
-                Statuses0 = (Statuses0)br.ReadUInt16();
-                //0x0010  2 bytes status_0; //statuses 0-7
-                Item_Index = br.ReadByte();
-                //0x0012  1 byte Used item index
-                Crit = br.ReadByte();
-                //0x0013  1 byte Crit increase
-                Statuses1 = (Statuses1)br.ReadUInt32();
-                //0x0014  4 bytes status_1; //statuses 8-39
-            }
-        }
-
-        /// <summary>
-        /// Duel (Zell limit break)
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Duel-%28Zell-limit-break%29"/>
-        /// <seealso cref="https://finalfantasy.fandom.com/wiki/Duel"/>
-        internal class Duel_Zell_limit_break
-        {
-            internal const int count = 10;
-            internal const int id = 22;
-            internal const int size = 32;
-            public override string ToString() => Name;
-            public FF8String Name { get; private set; }
-            public FF8String Description { get; private set; }
-            public Magic_ID MagicID { get; private set; }
-            public Attack_Type Attack_Type { get; private set; }
-            public byte Attack_Power { get; private set; }
-            public Attack_Flags Attack_Flags { get; private set; }
-            public byte Unknown0 { get; private set; }
-            public Target Target { get; private set; }
-            public byte Unknown1 { get; private set; }
-            public byte Hit_Count { get; private set; }
-            public Element Element { get; private set; }
-            public byte Element_Percent { get; private set; }
-            public byte Status_Attack { get; private set; }
-            public Buttons[] Button_Combo { get; private set; }
-            public Statuses0 Statuses0 { get; private set; }
-            public Statuses1 Statuses1 { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-                br.BaseStream.Seek(4, SeekOrigin.Current);
-                MagicID = (Magic_ID)br.ReadUInt16();
-                //0x0004  2 bytes Magic ID
-                Attack_Type = (Attack_Type)br.ReadByte();
-                //0x0006  1 byte Attack type
-                Attack_Power = br.ReadByte();
-                //0x0007  1 byte Attack power
-                Attack_Flags = (Attack_Flags)br.ReadByte();
-                //0x0008  1 byte Attack flags
-                Unknown0 = br.ReadByte();
-                //0x0009  1 byte Unknown
-                Target = (Target)br.ReadByte();
-                //0x000A  1 byte Target Info
-                Unknown1 = br.ReadByte();
-                //0x000B  1 bytes Unknown
-                Hit_Count = br.ReadByte();
-                //0x000C  1 byte Hit count
-                Element = (Element)br.ReadByte();
-                //0x000D  1 byte Element Attack
-                Element_Percent = br.ReadByte();
-                //0x000E  1 byte Element Attack %
-                Status_Attack = br.ReadByte();
-                //0x000F  1 byte Status Attack Enabler
-                Button_Combo = new Buttons[5];
-                for (int b = 0; b < Button_Combo.Length; b++)
-                {
-                    Button_Flags key = (Button_Flags)br.ReadUInt16();
-                    if (Input.Convert_Button.ContainsKey(key))
-                        Button_Combo[b] = Input.Convert_Button[key];
-                }
-                //0x0010  2 bytes Sequence Button 1
-                //0x0012  2 bytes Sequence Button 2
-                //0x0014  2 bytes Sequence Button 3
-                //0x0016  2 bytes Sequence Button 4
-                //0x0018  2 bytes Sequence Button 5
-                Statuses0 = (Statuses0)br.ReadUInt16();
-                //0x001A  2 bytes status_0; //statuses 0-7
-                Statuses1 = (Statuses1)br.ReadUInt32();
-                //0x001C  4 bytes status_1; //statuses 8-39
-            }
-        }
-        /// <summary>
-        /// Zell Limit Break Parameters
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Zell-limit-break-parameters"/>
-        internal class Zell_limit_break_parameters
-        {
-            internal const int count = 24;
-            internal const int id = 23;
-            internal const int size = 4;
-
-            public byte[] Moves { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Moves = br.ReadBytes(4);
-                //0x0000  1 byte[[StartMove 0
-                //0x0001  1 byte  Next Sequence 0_1
-                //0x0002  1 byte  Next Sequence 0_2
-                //0x0003  1 byte  Next Sequence 0_3
-            }
-        }
-        /// <summary>
-        /// Rinoa limit breaks (part 1)
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Rinoa-limit-breaks-%28part-1%29"/>
-        internal class Rinoa_limit_breaks_part_1
-        {
-            internal const int count = 2;
-            internal const int id = 24;
-            internal const int size = 8;
-
-            public override string ToString() => Name;
-            public FF8String Name { get; private set; }
-            public FF8String Description { get; private set; }
-            public BitArray Unknown0 { get; private set; }
-            public Target Target { get; private set; }
-            public byte AbilityID { get; private set; }
-            public byte Unknown1 { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2);
-                //0x0000	2 bytes Offset to name
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i * 2 + 1);
-                //0x0002	2 bytes Offset to description
-                br.BaseStream.Seek(4, SeekOrigin.Current);
-                Unknown0 = new BitArray(br.ReadBytes(1));
-                //0x0004  1 byte Unknown Flags
-                Target = (Target) br.ReadByte();
-                //0x0005  1 byte Target
-                AbilityID = br.ReadByte();
-                //0x0006  1 byte Ability data ID
-                Unknown1 = br.ReadByte();
-                //0x0007  1 byte Unknown / Unused
-            }
-        }
-
-        /// <summary>
-        /// Rinoa limit breaks (part 2)
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Rinoa-limit-breaks-%28part-2%29"/>
-        internal class Rinoa_limit_breaks_part_2
-        {
-            internal const int count = 33;
-            internal const int id = 25;
-            internal const int size = 20;
-
-            public override string ToString() => Name;
-            public FF8String Name { get; private set; }
-            public Magic_ID MagicID { get; private set; }
-            public Attack_Type Attack_Type { get; private set; }
-            public byte Attack_Power { get; private set; }
-            public Attack_Flags Attack_Flags { get; private set; }
-            public byte Unknown0 { get; private set; }
-            public Target Target { get; private set; }
-            public byte Unknown1 { get; private set; }
-            public byte Hit_Count { get; private set; }
-            public Element Element { get; private set; }
-            public byte Element_Percent { get; private set; }
-            public byte Status_Attack { get; private set; }
-            public Statuses0 Statuses0 { get; private set; }
-            public Statuses1 Statuses1 { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, i);
-                //0x0000	2 bytes Offset to name
-                br.BaseStream.Seek(2, SeekOrigin.Current);
-                MagicID = (Magic_ID)br.ReadUInt16();
-                //0x0002  2 bytes Magic ID
-                Attack_Type = (Attack_Type)br.ReadByte();
-                //0x0004  1 byte Attack type
-                Attack_Power= br.ReadByte();
-                //0x0005  1 byte Attack power
-                Attack_Flags = (Attack_Flags)br.ReadByte();
-                //0x0006  1 byte Attack flags
-                Unknown0 = br.ReadByte();
-                //0x0007  1 byte Unknown
-                Target = (Target)br.ReadByte();
-                //0x0008  1 byte Target info
-                Unknown1 = br.ReadByte();
-                //0x0009  1 byte Unknown
-                Hit_Count = br.ReadByte();
-                //0x000A  1 byte Hit Count
-                Element = (Element)br.ReadByte();
-                //0x000B  1 byte Element Attack
-                Element_Percent = br.ReadByte();
-                //0x000C  1 byte Element Attack %
-                Status_Attack=br.ReadByte();
-                //0x000D  1 byte Status Attack Enabler
-                Statuses0 = (Statuses0)br.ReadUInt16();
-                //0x000E  2 bytes status_0; //statuses 0-7
-                Statuses1 = (Statuses1)br.ReadUInt32();
-                //0x0010  4 bytes status_1; //statuses 8-39
-            }
-        }
-        /// <summary>
-        /// Slot Array Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Slot-array"/>
-        internal class Slot_array
-        {
-            internal const int count = 60;
-            internal const int id = 26;
-            internal const int size = 1;
-
-            public byte SlotID { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                SlotID = br.ReadByte();
-            }
-        }
-        /// <summary>
-        /// Slot Sets Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Selphie-limit-break-sets"/>
-        /// <seealso cref="https://finalfantasy.fandom.com/wiki/Slots_(ability_type)#Final_Fantasy_VIII"/>
-        internal class Selphie_limit_break_sets
-        {
-            internal const int count = 16;
-            internal const int id = 27;
-            internal const int size = 16;
-            internal Slot[] Slots { get; private set; }
-            internal void Read(BinaryReader br, int i)
-            {
-                Slots = new Slot[8];
-                for (int s=0; s< 8; s++)
-                {
-                    Slots[s].Read(br, s);
-                }
-            }
-        }
-        /// <summary>
-        /// Slot Magic Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Selphie-limit-break-sets"/>
-        internal struct Slot
-        {
-            internal byte MagicID { get; private set; }
-            internal byte Count { get; private set; }
-
-            internal void Read(BinaryReader br, int i=0)
-            {
-                MagicID = br.ReadByte();
-                Count = br.ReadByte();
-            }
-        }
-
-        [Flags]
-        enum StatFlags
-        {
-            None = 0x00,
-            STR = 0x01,
-            VIT = 0x02,
-            MAG = 0x04,
-            SPR = 0x08,
-            SPD = 0x10,
-        }
-        [Flags]
-        enum Quanity
-        {
-            //0% = 0x00,
-            //6.25% = 0x01,
-            //12.50% = 0x02,
-            //25% = 0x04,
-            //50% = 0x08,
-            //100% = 0x10,
-            _0f = 0x00,
-            _0625f = 0x01,
-            _1250f = 0x02,
-            _25f = 0x04,
-            _50f = 0x08,
-            _1f = 0x10,
-        }
-        /// <summary>
-        /// Devour Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Devour"/>
-        internal class Devour
-        {
-            internal const int count = 16;
-            internal const int id = 28;
-            internal const int size = 12;
-            public override string ToString() => Description;
-
-            public FF8String Description { get; private set; }
-            public float Amount { get; private set; }
-            /// <summary>
-            /// True for heal, False for damage
-            /// </summary>
-            public bool DMGorHEAL { get; private set; }
-            public Statuses1 Statuses1 { get; private set; }
-            public Statuses0 Statuses0 { get; private set; }
-            private StatFlags StatFlags { get; set; }
-            public byte HP { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Description = Memory.Strings.Read(Strings.FileID.KERNEL, id, i);
-                br.BaseStream.Seek(2, SeekOrigin.Current);
-                //0x0000  2 bytes Offset to devour description
-                DMGorHEAL = br.ReadByte() == 0x1E ? true : false;
-                //0x0002  1 byte Damage or heal HP and Status
-
-                //0x1E - Cure
-                //0x1F - Damage
-                var val = (Quanity)br.ReadByte();
-                Amount = 0f;
-                if ((val & Quanity._0625f)!=0) Amount += .0625f;
-                if ((val & Quanity._1250f) != 0) Amount += .1250f;
-                if ((val & Quanity._1f) != 0) Amount += 1f;
-                if ((val & Quanity._25f) != 0) Amount += .25f;
-                if ((val & Quanity._50f) != 0) Amount += .50f;
-                //0x0003  1 byte HP Heal / DMG Quantity Flag
-
-                //0x00 - 0 %
-                //0x01 - 6.25 %
-                //0x02 - 12.50 %
-                //0x04 - 25 %
-                //0x08 - 50 %
-                //0x10 - 100 %
-                Statuses1 = (Statuses1)br.ReadUInt32();
-                //0x0004  4 bytes status_1; //statuses 8-39
-                Statuses0 = (Statuses0)br.ReadUInt16();
-                //0x0008  2 bytes status_0; //statuses 0-7
-
-                StatFlags = (StatFlags)br.ReadByte();
-                //0x000A  1 byte Raised Stat Flag
-
-                //0x00 - None
-                //0x01 - STR
-                //0x02 - VIT
-                //0x04 - MAG
-                //0x08 - SPR
-                //0x10 - SPD
-                HP = br.ReadByte();
-                //0x000B  1 byte Raised Stat HP Quantity
-            }
-        }
-        /// <summary>
-        /// Misc Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Misc-section"/>
-        internal class Misc_section
-        {
-            internal const int count = 1;
-            internal const int id = 29;
-
-            public byte[] Status_Timers { get; private set; }
-            public byte ATB_Speed_Multiplier { get; private set; }
-            public byte Dead_Timer { get; private set; }
-            public byte[] Status_Limit_Effects { get; private set; }
-            public byte[] Duel_Timers_and_Start_Moves { get; private set; }
-            public byte[] Shot_Timers { get; private set; }
-
-            internal void Read(BinaryReader br, int i)
-            {
-                Status_Timers = br.ReadBytes(14);
-                ATB_Speed_Multiplier = br.ReadByte();
-                Dead_Timer = br.ReadByte();
-                Status_Limit_Effects = br.ReadBytes(32);
-                Duel_Timers_and_Start_Moves = br.ReadBytes(8);
-                Shot_Timers = br.ReadBytes(4);
-            }
-        }
-        /// <summary>
-        /// Misc Text Pointers Data
-        /// </summary>
-        /// <see cref="https://github.com/alexfilth/doomtrain/wiki/Misc-text-pointers"/>
-        internal class Misc_text_pointers
-        {
-            internal const int count = 128;
-            internal const int id = 30;
-            internal const int size = 2;
-
-            public override string ToString() => Value;
-            public static explicit operator FF8String(Misc_text_pointers v) => v.Value;
-            internal FF8String Value { get; private set; }
-
-            //0x0000	2 bytes Offset to item name
-
-            internal void Read(int i)
-            {
-                Value = Memory.Strings.Read(Strings.FileID.KERNEL, id, i);
-                //0x0000	2 bytes Offset to item name
-            }
-        }
+}
     }
 }
